requirements.txt
opencv-python>=4.8
mediapipe>=0.10
numpy>=1.24
scipy>=1.11
Pillow>=10.0
